+++ site-config/anvil-cmg/prod/config.ts
@@ -1,8 +1,6 @@
-import { EntityConfig } from "@databiosphere/findable-ui/lib/config/entities";
 import { GIT_HUB_REPO_URL } from "../../common/constants";
 import { SiteConfig } from "../../common/entities";
 import { makeConfig } from "../dev/config";
-import { downloadColumn } from "../dev/index/filesEntityConfig";
 import { authenticationConfig } from "./authentication/authentication";
 
 const config: SiteConfig = {
@@ -11,17 +9,12 @@ const config: SiteConfig = {
     "https://anvilproject.org",
     "https://service.explore.anvilproject.org",
     GIT_HUB_REPO_URL,
-    "anvil12"
+    "anvil13"
   ),
   exportToTerraUrl: "https://anvil.terra.bio/",
 };
 
 config.authentication = authenticationConfig;
-
-const filesEntityConfig = config.entities.find(
-  (c) => c.apiPath == "index/files"
-) as EntityConfig;
-filesEntityConfig.list.columns.splice(0, 0, downloadColumn);
 
 // Update gtmAuth for the prod environment lookup.
 if (config.analytics) {