+++ CHANGELOG.md
@@ -1,5 +1,43 @@
 # Changelog
 
+## [2.27.0](https://github.com/DataBiosphere/data-browser/compare/v2.26.1...v2.27.0) (2026-03-12)
+
+
+### Features
+
+* update anvil cmg auth config to use ecm endpoint ([#4694](https://github.com/DataBiosphere/data-browser/issues/4694)) ([#4696](https://github.com/DataBiosphere/data-browser/issues/4696)) ([ed2262a](https://github.com/DataBiosphere/data-browser/commit/ed2262aecda61eb9377c950ae206d05c05e684b1))
+
+
+### Chores
+
+* upgrade findable-ui to 49.4.1 ([#4693](https://github.com/DataBiosphere/data-browser/issues/4693)) ([e7176c6](https://github.com/DataBiosphere/data-browser/commit/e7176c6b420de7515b4d6563fb236cbd6b2a4924))
+
+## [2.26.1](https://github.com/DataBiosphere/data-browser/compare/v2.26.0...v2.26.1) (2026-03-05)
+
+
+### Chores
+
+* update hca-dcp catalog to dcp57 ([#4688](https://github.com/DataBiosphere/data-browser/issues/4688)) ([b69ac23](https://github.com/DataBiosphere/data-browser/commit/b69ac239a8db165fa295d915c68285f38d38229f))
+* update hca-dcp catalog to dcp57 [#4685](https://github.com/DataBiosphere/data-browser/issues/4685) ([b69ac23](https://github.com/DataBiosphere/data-browser/commit/b69ac239a8db165fa295d915c68285f38d38229f))
+
+## [2.26.0](https://github.com/DataBiosphere/data-browser/compare/v2.25.0...v2.26.0) (2026-02-27)
+
+
+### Features
+
+* add conditional file download for anvil-cmg ([#4635](https://github.com/DataBiosphere/data-browser/issues/4635)) ([#4679](https://github.com/DataBiosphere/data-browser/issues/4679)) ([90b3cff](https://github.com/DataBiosphere/data-browser/commit/90b3cff6acad1bf1b8175e1ab72bc9dbf74f28e0))
+
+
+### Bug Fixes
+
+* add deadline date to RAS banner ([#4681](https://github.com/DataBiosphere/data-browser/issues/4681)) ([#4682](https://github.com/DataBiosphere/data-browser/issues/4682)) ([8e4aa01](https://github.com/DataBiosphere/data-browser/commit/8e4aa012275c6b04b383360c55296101935e1654))
+
+
+### Chores
+
+* update anvil-cmg catalog to anvil13 ([#4683](https://github.com/DataBiosphere/data-browser/issues/4683)) ([7699401](https://github.com/DataBiosphere/data-browser/commit/76994014f6d2c0bce3e30c0ebf4c53e7cb4cb5c5))
+* update anvil-cmg catalog to anvil13 ([#4684](https://github.com/DataBiosphere/data-browser/issues/4684)) ([7699401](https://github.com/DataBiosphere/data-browser/commit/76994014f6d2c0bce3e30c0ebf4c53e7cb4cb5c5))
+
 ## [2.25.0](https://github.com/DataBiosphere/data-browser/compare/v2.24.0...v2.25.0) (2026-02-19)